absenteeism/requirements.txt
pandas
sf-hamilton
metaflow
numpy
sklearn
imbalanced-learn
skrebate
matplotlib
xgboost
torch 
skorch